requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clr-loader==0.1.7
cycler==0.11.0
dataframe-image==0.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.15.3
fonttools==4.32.0
fpdf==1.7.2
frozenlist==1.3.0
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
kaleido==0.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.0
matplotlib==3.5.1
mistune==0.8.4
multidict==6.0.2
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
numpy==1.22.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
Pillow==9.1.0
plotly==5.6.0
pycparser==2.21
pyodbc==4.0.32
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pythonnet==3.0.0a2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2
tenacity==8.0.1
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
webencodings==0.5.1
yarl==1.7.2
